+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 Changes:
 
+- Add support for Ledger `signRaw`
 - Upgrade dependencies to latest stable versions
 - Use `stringify/import.meta.url` helpers in specs
+
 
 ## 12.3.2 Jun 12, 2023